utf8: add simple benchmarks for ASCII-only text
Probably the most important case is when the text being verified
is (mostly) ASCII. Create a simple benchmark that takes 500MB of
a NULL-terminated ASCII string, and verifies that it is valid UTF8.

For the sake of comparison, compare it with a trivial implementation
(essentially the algorithm used by glib) and compare it with strlen().
The former should be a worst-case that we should beat, and the latter
a best case that we can't really. We observe a 350% speed-up compared
with the trivial implementation.

On my machine the numbers are:
UTF-8 verify ASCII string (trivial): 1642 MB/s
UTF-8 verify ASCII string: 7397 MB/s
strlen() of ASCII string: 7837 MB/s

Signed-off-by: Tom Gundersen <teg@jklm.no>